Cleanup duplicated functions
