+++ src/sources/hex8.rs
@@ -1,6 +1,36 @@
 use crate::MU0_4PI;
-use crate::math::{cross, distance};
-use crate::sources::{edge_csys, edge_integral, transform};
+use crate::math::{cross, distance, dot3, mag3, unit_vector};
+use crate::sources::element::edge_integral;
+// transform a global xyz position into a local coordinate system
+fn transform(g: &[f64; 3], xhat: &[f64; 3], yhat: &[f64; 3], zhat: &[f64; 3]) -> [f64; 3] {
+    [dot3(xhat, g), dot3(yhat, g), dot3(zhat, g)]
+}
+
+// Compute the unit vectors of an edge csys given 3 nodes
+// Nodes should be in right hand rule order
+// middle node is the 'anchor', xhat points from node 1 to node 0
+// yhat points away from the center of the face
+fn edge_csys(x: &[f64; 3], y: &[f64; 3], z: &[f64; 3]) -> ([f64; 3], [f64; 3], [f64; 3]) {
+    // first edge defines x'-axis
+    let xp_hat = unit_vector(&[x[1], y[1], z[1]], &[x[0], y[0], z[0]]);
+
+    // second edge defines the next in plane direction to find the out of plane z
+    // it is not returned because it is not guaranteed to be orthogonal to x and z
+    let yp_hat_temp = unit_vector(&[x[2], y[2], z[2]], &[x[1], y[1], z[1]]);
+
+    // unit vector normal to surface
+    let mut zp_hat = [0.0; 3];
+    cross(&xp_hat, &yp_hat_temp, &mut zp_hat);
+    let norm = mag3(zp_hat[0], zp_hat[1], zp_hat[2]);
+    zp_hat[0] /= norm;
+    zp_hat[1] /= norm;
+    zp_hat[2] /= norm;
+
+    // other in-plane direction, not necessarily aligned with an edge
+    let mut yp_hat = [0.0; 3];
+    cross(&xp_hat, &zp_hat, &mut yp_hat);
+    (xp_hat, yp_hat, zp_hat)
+}
 
 /// Compute the magnetic flux density generated by an arbitrary hexahedron in 3D space
 ///
